+++ crankshaft-monitor/CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-* Added support for the `ImagePull{Started,Failed,Finished}` events ([#](https://github.com/stjude-rust-labs/crankshaft/pull/)).
+* Added support for the `ImagePull{Started,Failed,Finished}` events ([#82](https://github.com/stjude-rust-labs/crankshaft/pull/82)).
 
 ## 0.2.0 - 02-11-2026